feat(canvas): add useCanvas api
