+++ src/types/index.ts
@@ -23,6 +23,8 @@ import {
 import { UseEventHandlers } from './use-event-handlers';
 import { DesignModeBuiltIns, RunModeBuiltIns } from './builtins';
 import { UseDataSourceApi } from './properties';
+import { UseCanvasApi } from './canvas';
+import { UseComponentStateApi, UseComponentSubscriberApi } from './observer';
 
 export * from './component-key';
 export * from './container-context-data';
@@ -34,6 +36,8 @@ export * from './trans';
 export * from './use-event-handlers';
 export * from './model';
 export * from './properties';
+export * from './canvas';
+export * from './observer';
 
 export { DesignModeBuiltIns, RunModeBuiltIns };
 
@@ -54,6 +58,7 @@ export type CommonSDK = {
   transMetaVal2PureVal: typeof transMetaVal2PureVal;
   transObjectKeys: typeof transObjectKeys;
   useDynamicValue: typeof useDynamicValue;
+  useCanvas: UseCanvasApi;
 };
 
 export type DesignModeSDKInjection = {
@@ -73,6 +78,8 @@ export type RunModeSDKInjection = {
   modifyRecordAttribute: ModifyRecordAttributeMethod;
   fetchRecordsByCondition: FetchRecordsByConditionMethod;
   fetchByDataSource: FetchByDataSourceMethod;
+  useComponentState: UseComponentStateApi;
+  useComponentSubscriber: UseComponentSubscriberApi;
 };
 
 export type RunModeSDK = CommonSDK & RunModeSDKInjection;
